pyproject.toml
[project]
name = "autonomous-analyst"
version = "0.1.0"
description = "Production-grade autonomous analyst swarm"
readme = "READMEGUEY.md"
requires-python = ">=3.11"
dependencies = [
    "altair<5",
    "certifi>=2026.2.25",
    "fastapi>=0.135.1",
    "fpdf2>=2.8.7",
    "langchain-anthropic>=1.3.4",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.11",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.1.2",
    "matplotlib>=3.10.8",
    "mcp>=1.26.0",
    "numpy>=2.4.3",
    "pandas<3.0.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.48",
    "sqlalchemy-cockroachdb>=2.0.3",
    "streamlit>=1.24.0",
    "tavily-python>=0.7.23",
    "uvicorn>=0.41.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
managed = true
package = false
